Trainer cli related tests moved to test_trainer_cli.py
